anime_recommender/backend/requirements.txt
flask
flask-cors
pandas
numpy
scikit-learn
scipy
rapidfuzz
python-dotenv
google-genai
